Crate l2math

Low Level Math

A port of MUSL's Math library to Rust for no_std environments.

Disclaimer

It is exceptionally rare that you should use this library directly. It is intended to be used by other libraries that require low level math functionality. If you are looking for a math library, you should probably be using libtrig.

Macros

Σ

Sigma Σ macro

Constants

E

Euler’s number (e)

EF

Euler’s number (e) as a 32-bit float

EGAMMA

The Euler-Mascheroni constant (γ)

EGAMMAF

The Euler-Mascheroni constant (γ) as a 32-bit float

FRAC_1_PI

1/π

FRAC_1_PIF

1/π as a 32-bit float

FRAC_1_SQRT_2

1/sqrt(2)

FRAC_1_SQRT_2F

1/sqrt(2) as a 32-bit float

FRAC_1_SQRT_3

1/sqrt(3)

FRAC_1_SQRT_3F

1/sqrt(3) as a 32-bit float

FRAC_1_SQRT_PI

1/sqrt(π)

FRAC_1_SQRT_PIF

1/sqrt(π) as a 32-bit float

FRAC_2_PI

2/π

FRAC_2_PIF

2/π as a 32-bit float

FRAC_2_SQRT_PI

2/sqrt(π)

FRAC_2_SQRT_PIF

2/sqrt(π) as a 32-bit float

FRAC_PI_2

π/2

FRAC_PI_2F

π/2 as a 32-bit float

FRAC_PI_3

π/3

FRAC_PI_3F

π/3 as a 32-bit float

FRAC_PI_4

π/4

FRAC_PI_4F

π/4 as a 32-bit float

FRAC_PI_6

π/6

FRAC_PI_6F

π/6 as a 32-bit float

FRAC_PI_8

π/8

FRAC_PI_8F

π/8 as a 32-bit float

LN_2

ln(2)

LN_2F

ln(2) as a 32-bit float

LN_10

ln(10)

LN_10F

ln(10) as a 32-bit float

LOG2_10

log2(10)

LOG2_10F

log2(10) as a 32-bit float

LOG2_E

log2(e)

LOG2_EF

log2(e) as a 32-bit float

LOG10_2

log10(2)

LOG10_2F

log10(2) as a 32-bit float

LOG10_E

log10(e)

LOG10_EF

log10(e) as a 32-bit float

PHI

The golden ratio (φ)

PHIF

The golden ratio (φ) as a 32-bit float

PI

Archimedes’ constant (π)

PIF

Archimedes’ constant (π) as a 32-bit float

SQRT_2

sqrt(2)

SQRT_2F

sqrt(2) as a 32-bit float

SQRT_3

sqrt(3)

SQRT_3F

sqrt(3) as a 32-bit float

TAU

The full circle constant (τ)

TAUF

The full circle constant (τ) as a 32-bit float

Functions

acos

Arccosine

acosf

Arccosine

acosh

Inverse hyperbolic cosine

acoshf

Inverse hyperbolic cosine

asin

Arcsine

asinf

Arcsine

asinh

Inverse hyperbolic sine

asinhf

Inverse hyperbolic sine

atan

Arctangent

atan2

Arctangent of y/x

atan2f

Arctangent of y/x

atanf

Arctangent

atanh

Inverse hyperbolic tangent

atanhf

Inverse hyperbolic tangent

cbrt

Computes the cube root of the argument.

cbrtf

Cube root

ceil

Ceil

ceilf

Ceil

copysign

Sign of Y, magnitude of X

copysignf

Sign of Y, magnitude of X

cos

Cosine

cosf

Cosine

cosh

Hyperbolic cosine

coshf

Hyperbolic cosine

erf

Error function

erfc

Complementary error function

erfcf

Complementary error function

erff

Error function

exp

Exponential, base e

exp2

Exponential, base 2

exp2f

Exponential, base 2

exp10

Exponential, base 10

exp10f

Exponential, base 10

expf

Exponential, base e

expm1

Exponential, base e, of x-1

expm1f

Exponential, base e, of x-1

fabs

Absolute value (magnitude)

fabsf

Absolute value (magnitude)

fdim

Positive difference

fdimf

Positive difference

floor

Floor

floorf

Floor

fma

Floating multiply add

fmaf

Fused multiply-add: Compute x * y + z with a single rounding error.

fmax

Returns the maximum of the two arguments, signaling NaNs if either argument is a signaling NaN.

fmaxf

Returns the maximum of the two arguments, signaling NaNs if either argument is a signaling NaN.

fmin

Returns the minimum of two numbers.

fminf

Returns the minimum of two numbers.

fmod

Remainder of floating point division

fmodf

Remainder of floating point division

frexp

Breaks the number into a normalized fraction and a base-2 exponent

frexpf

Breaks the number into a normalized fraction and a base-2 exponent

hypot

Calculate the length of the hypotenuse of a right-angle triangle given legs of length x and y.

hypotf

Calculate the length of the hypotenuse of a right-angle triangle given legs of length x and y.

ilogb

Get exponent of floating point value

ilogbf

Get exponent of floating point value

j0

Bessel function of the first kind of order zero

j0f

Bessel function of the first kind of order zero

j1

Bessel function of the first kind of order one

j1f

Bessel function of the first kind of order one

jn

Bessel function of the first kind of order zero of x.

jnf

Bessel function of the first kind of order zero

ldexp

Returns x * 2n.

ldexpf

Returns x * 2n.

lgamma

Natural logarithm of gamma function

lgamma_r

Natural logarithm of gamma function

lgammaf

Natural logarithm of gamma function

lgammaf_r

Natural logarithm of gamma function

ln

Return the natural logarithm of x.

ln1p

Return the natural logarithm of 1+x.

ln1pf

Return the natural logarithm of 1+x.

lnf

Return the natural logarithm of x.

log

Returns the logarithm of x.

log2

Returns the base 2 logarithm of x.

log2f

Returns the base 2 logarithm of x.

log10

Returns the base 10 logarithm of x.

log10f

Returns the base 10 logarithm of x.

logf

Returns the logarithm of x

modf

Breaks the given number into an integral and a fractional part.

modff

Breaks the given number into an integral and a fractional part.

nextafter

Returns the next representable floating-point value following x in the direction of y.

nextafterf

Returns the next representable floating-point value following x in the direction of y.

pow

Returns x raised to the power y.

powf

Returns x raised to the power y.

remainder

Returns the remainder of x/y.

remainderf

Returns the remainder of x/y.

remquo

Return the remainder and part of the quotient of x and y.

remquof

Return the remainder and part of the quotient of x and y.

rint

Round to nearest integer, rounding halfway cases away from zero.

rintf

Round to nearest integer, rounding halfway cases away from zero.

round

Rounds x to the nearest integer in the direction of the current rounding mode.

roundf

Rounds x to the nearest integer in the direction of the current rounding mode.

scalbn

Returns x * 2^n

scalbnf

Returns x * 2^n

sin

Returns the sine function of x.

sincos

Simultaneously computes the sine and cosine of the argument x.

sincosf

Simultaneously computes the sine and cosine of the argument x.

sinf

Returns the sine of the argument x.

sinh

Returns the hyperbolic sine of x.

sinhf

Returns the hyperbolic sine of x.

sqrt

Returns the square root of x.

sqrtf

Returns the square root of x.

tan

Returns tangent of x.

tanf

Returns tangent of x.

tanh

Returns the hyperbolic tangent of x.

tanhf

Returns the hyperbolic tangent of x.

tgamma

Returns the gamma function of x.

tgammaf

Returns the gamma function of x.

trunc

Returns the integer part of self. This means that non-integer numbers are always truncated towards zero.

truncf

Returns the integer part of self. This means that non-integer numbers are always truncated towards zero.

ulp

Returns the value of the least significant bit of the given floating point number.

y0

Bessel function of the second kind of order zero

y0f

Bessel function of the second kind of order zero

y1

Bessel function of the second kind of order one

y1f

Bessel function of the second kind of order one

yn

Bessel function of the second kind of order zero of x.

ynf

Bessel function of the second kind of order zero

Type Aliases

Degree32

Degrees (32bit)

Degree64

Degrees (64bit)

Float32

A floating point number (32bit)

Float64

A floating point number (64bit)

Int

A signed integer (32bit)

Radian32

Radians (32bit)

Radian64

Radians (64bit)